refactor: move token usage to dedicated Usage tab in settings
Extract TokenUsageSection from ModelManager into its own UsageSettings
component and give it a dedicated "Usage" tab in the settings sidebar.
This clearly separates read-only analytics from editable model settings.
